[SDK] Added a processing delay config variable.
[rpc]
processingdelay=0 (default) [milliseconds]

This helps with lowering cpu usage due to software pushing out color updates as fast as the receiver accepts them.
If you want to keep your cpu usage low i encourage you to set it to 1 millisecond. But keep in mind that doing that will add a tiny delay.
If after that change you experience issues with the application, like lower framerates or dropped frames remove the delay as the sending application might use it in a timing critical part of its main logic.